requirements.txt
# Core
python-dotenv>=0.19.0
pandas>=1.3.0
numpy>=1.21.0
python-dateutil>=2.8.2

# Web Framework
flask>=2.0.1
flask-login>=0.5.0
flask-wtf>=1.0.0

# Security
cryptography>=3.4.7
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Threat Intelligence
requests>=2.26.0
vt-py>=0.12.0
otx-misp>=1.0.0
abuseipdb>=0.1.0

# Machine Learning
scikit-learn>=1.0.0
joblib>=1.1.0

# Email Processing
easyimap>=0.9.9
email-validator>=1.1.3

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1

# Development
black>=21.12b0
flake8>=4.0.1
mypy>=0.930

# Documentation
sphinx>=4.2.0
sphinx-rtd-theme>=1.0.0
